用户相关接口初版

+++ trade-data/src/main/java/com/russellzhou/trade/data/entity/UserInfo.java
@@ -44,7 +44,7 @@ public class UserInfo {
     /**
      *   是否删除
      */
-    private Byte isDel;
+    private Integer isDel;
 
     /**
      *   创建时间
